style: clippy --fix 机械自动修复（批次A，约110条）
覆盖 needless_borrow 系 / collapsible_if / excessive_precision（f32 字面量
截断，零语义差异）/ explicit_auto_deref / identity_op 等机械可修类别，
全量 3300+ 测试通过。行为零变更。

+++ src/data/transforms/random_resized_crop.rs
@@ -156,7 +156,7 @@ fn get_crop_params(
 
     // Fallback: 中心裁切，保持宽高比
     let in_ratio = w as f64 / h as f64;
-    let mid_ratio = ((ratio.0 * ratio.1) as f64).sqrt();
+    let mid_ratio = (ratio.0 * ratio.1).sqrt();
 
     let (crop_h, crop_w) = if in_ratio < mid_ratio {
         // 原图太窄，以宽度为准
